+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD
  * [preloader] Parse klog flags.
+* [CTFE] Add a /log.v3.json endpoint to help satisfy a requirement of the Chrome CT Log Policy by @robstradling in https://github.com/google/certificate-transparency-go/pull/1703
 
 ## v1.3.2